Player updated
Sets up it's pixels correctly. Added consts variables for the different states of the player (attacking, roaming) and the change in bounds for these states. Added functions that start and stop attacking, which changes the bounds and colors of the player (single, green LED for roaming, 5 blues LED's for attacking). Also make pixels array public so that it can be updated in main