Drawing, Attacking and Patrolling
Got drawing semi-working... small bug that still needs to be fixed (should be easy if we do a check for the attack bool).

Attack animation also looks good. Just needs to kill stuff now :D

Tweaked the patrolling a bit so it is easier to see what will happen to the enemy.