Updates (needs testing)
Added check for player colliding with left side of board - player gets stuck so that it doesn't keep trying to move when it is at the edge of the board. Added collision check with patroller and lava. If the player dies or the patroller dies, their drawEntity stops drawing any pixels.